bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.25.0",
    "aiohttp>=3.13.3",
    "python-dotenv>=1.2.1",
]
